test: add traceability guidance RED inventory
