+++ packages/tests/repo/rallar-authoritative-mutation-guidance-integrity.test.ts
@@ -1,0 +1,396 @@
+import { existsSync, readFileSync } from 'node:fs';
+import path from 'node:path';
+import { describe, expect, it } from 'vitest';
+
+const repoRoot = process.cwd();
+const canonicalServiceWritingPath =
+  '.agents/skills/rallar-code-writing/references/convergent-service-writing.md';
+const codeProducingSpecialistSkills = [
+  'building-rallar-apps',
+  'performance-analysis',
+  'rallar-ai',
+  'rallar-games',
+  'rallar-hetzner-ops',
+  'rallar-platform',
+  'rallar-realtime',
+] as const;
+const appInboxGuidanceVocabulary = [
+  'AppInbox is mandatory for incoming database mutations',
+  'service write receives the transaction',
+  'ResourceInboxRepository',
+  'APP_OUTBOX',
+  'WS_OUTBOX',
+  '20 total processing attempts',
+  'mandatory fields by default',
+] as const;
+const currentAppInboxGuidancePaths = [
+  canonicalServiceWritingPath,
+  'packages/shared-server/architecture.md',
+  'packages/shared-server/rallar-server-repositories.md',
+  'packages/shared-server/rallar-server-repositories-improvements.md',
+  'apps/api-v1/README.md',
+  'docs/rallar-api-reference.md',
+  'docs/rallar-convergent-state-and-rtc-topology.md',
+  'docs/rallar-crdt-guide.md',
+  'docs/rallar-crdt-production-hardening-runbook.md',
+  'docs/README.md',
+] as const;
+const coreConvergentWriteGuidancePaths = [canonicalServiceWritingPath] as const;
+const canonicalSnapshotOrderingGuidancePaths = [
+  canonicalServiceWritingPath,
+  'docs/rallar-convergent-state-and-rtc-topology.md',
+] as const;
+const postCommitAudienceGuidancePaths = [
+  'apps/api-v1/README.md',
+  'packages/shared-server/architecture.md',
+  'packages/shared-server/rallar-server-repositories.md',
+  'docs/rallar-api-reference.md',
+  'docs/rallar-convergent-state-and-rtc-topology.md',
+] as const;
+const repositoryReadGuidancePaths = [
+  'apps/api-v1/README.md',
+  'docs/README.md',
+  'docs/rallar-api-reference.md',
+  'docs/rallar-crdt-guide.md',
+  'docs/rallar-crdt-production-hardening-runbook.md',
+] as const;
+const downstreamQueueGuidancePaths = [
+  'packages/shared-server/architecture.md',
+  'packages/shared-server/rallar-server-repositories.md',
+  'packages/shared-server/rallar-server-repositories-improvements.md',
+  'docs/rallar-convergent-state-and-rtc-topology.md',
+] as const;
+const mediumScaleRequirements = [
+  'npm run test:api-v1:black-box:postgres:medium-scale',
+  '100 independently authenticated clients',
+  'five groups',
+  'two Postgres-backed API processes',
+  '10 client lanes plus 5 control lanes',
+] as const;
+const performanceGateRequirements = [
+  'npm run perf:api-v1:state-write',
+  'node scripts/perf/compare-api-v1-state-write-results.mjs',
+  'comparative result gate',
+] as const;
+
+describe('authoritative mutation guidance integrity', () => {
+  it('routes composition and validation from the specialist skills', () => {
+    const agents = readRepo('AGENTS.md');
+    const convergenceArchitecture = readRepo('docs/rallar-convergent-state-and-rtc-topology.md');
+    const platform = readRepo('.agents/skills/rallar-platform/SKILL.md');
+    const codeWriting = readRepo('.agents/skills/rallar-code-writing/SKILL.md');
+    const repoCodeStyle = readRepo(
+      '.agents/skills/rallar-code-writing/references/repo-code-style.md',
+    );
+    const serviceWriting = readRepo(canonicalServiceWritingPath);
+    const games = readRepo('.agents/skills/rallar-games/SKILL.md');
+    const realtime = readRepo('.agents/skills/rallar-realtime/SKILL.md');
+    const testing = readRepo('.agents/skills/rallar-testing/SKILL.md');
+    const testCommands = readRepo('.agents/skills/rallar-testing/references/test-commands.md');
+    const packageJson = readJson('package.json') as {
+      scripts?: Readonly<Record<string, string>>;
+    };
+
+    expect(platform).toContain('building-rallar-apps');
+    expect(games).toContain('building-rallar-apps');
+    expect(games).toContain('Use the `rallar-testing` skill');
+    expect(realtime).toContain('building-rallar-apps');
+    expectAll(games, ['message.raw.targets', 'realtime payload', 'roomRef']);
+    expectAll(realtime, ['message.raw.targets', 'realtime payload', 'roomRef']);
+    expect(platform).toContain('Required fields are the default');
+    expect(platform).toMatch(/absence is a meaningful domain\s+state/);
+    expect(platform).toMatch(/explicitly ask\s+the human/);
+    expectAllNormalized(platform, [
+      canonicalServiceWritingPath,
+      'Platform-specific decisions remain here',
+      'authoritative persisted, replicated, queued, event, snapshot, and response',
+    ]);
+    expect(platform).not.toContain('optimistic compare-and-set writes with bounded retries');
+    expect(platform).not.toContain('Queue locks are coordination-only');
+    expect(platform).not.toContain('Strong contracts enable permissive convergence');
+    expectAll(codeWriting, [
+      'Always read `references/repo-code-style.md`',
+      'references/convergent-service-writing.md',
+      'Prefer required contracts and explicit value flow',
+      'compatibility fallback requires explicit human approval',
+    ]);
+    expectAllNormalized(serviceWriting, [
+      'expected-revision compare-and-set',
+      'authorization, policy, capacity, lifecycle, invariants',
+      'canonical key, decoded identity, complete mandatory shape',
+      'Corrupt authoritative reads fail closed',
+    ]);
+    const performance = readRepo('.agents/skills/performance-analysis/SKILL.md');
+    expectAll(performance, [
+      'production receipts',
+      'Synthetic post-call evidence cannot',
+      'Synthetic prerequisite or non-invocation evidence must link to an earlier same-subject predecessor with real production exhaustion',
+      'must not weaken candidate validation',
+    ]);
+    expectAll(repoCodeStyle, ['discriminated union']);
+    expectAllNormalized(serviceWriting, [
+      'create with conditional insert or `insertIfAbsent`',
+      'update with expected-revision compare-and-set or `upsertIfRevision`',
+      'delete or expire with expected-revision conditional delete or `deleteIfRevision`',
+      'a stale expiry observation must not delete or overwrite a refreshed value',
+      'A conflict starts a fresh `read`',
+    ]);
+    expectAllNormalized(realtime, [
+      canonicalServiceWritingPath,
+      'Presence summaries are optimistic materialized views, not authority',
+    ]);
+    expect(realtime).not.toContain('compare-and-set writes with bounded retries');
+    expect(realtime).not.toContain('Queue locks are coordination-only');
+    expect(realtime).not.toContain('Prefer optimistic reconciliation for replicated state');
+    expect(realtime).not.toContain('prove overlapping conflicts, rebasing, retry exhaustion');
+    expectAll(agents, [canonicalServiceWritingPath, 'functional core', 'stateful shell']);
+    expectAll(convergenceArchitecture, [
+      'Implemented Convergent Database Writes',
+      'conditional insert',
+      'expected-revision compare-and-set',
+      'expected-revision conditional delete',
+    ]);
+    expect(testing).toContain('npm run test:repo-governance');
+    expect(testCommands).toContain('npm run test:repo-governance');
+    expectAll(packageJson.scripts?.['test:repo-governance'] ?? '', [
+      'packages/tests/repo/rallar-authoritative-mutation-guidance-integrity.test.ts',
+      'packages/tests/repo/rallar-group-state-owner-integrity.test.ts',
+      'packages/tests/repo/rallar-skill-app-examples-integrity.test.ts',
+      'packages/tests/repo/rallar-skill-plugin-publication-integrity.test.ts',
+      'packages/tests/repo/repo-code-style-authority-integrity.test.ts',
+      'packages/tests/repo/repo-code-style-checker-integrity.test.ts',
+      'packages/tests/repo/repo-code-style-review-evidence-integrity.test.ts',
+      'packages/tests/repo/repo-style-structural-lineage-provenance.test.ts',
+      'packages/tests/repo/repo-style-check.test.ts',
+      'packages/tests/repo/repo-style-layout-rules.test.ts',
+      'packages/tests/repo/repo-style-changed-check.test.ts',
+      'packages/tests/rallar-black-box/rallar-testing-skill.test.ts',
+    ]);
+  });
+
+  it('routes TypeScript work through one code-writing skill and one service doctrine', () => {
+    const agents = readRepo('AGENTS.md');
+    const codeWriting = readRepo('.agents/skills/rallar-code-writing/SKILL.md');
+    const serviceWriting = readRepo(canonicalServiceWritingPath);
+    const testing = readRepo('.agents/skills/rallar-testing/SKILL.md');
+    const testCommands = readRepo('.agents/skills/rallar-testing/references/test-commands.md');
+
+    expect(existsSync(path.join(repoRoot, canonicalServiceWritingPath))).toBe(true);
+    expectAllNormalized(agents, [
+      canonicalServiceWritingPath,
+      'functional core',
+      'stateful shell',
+      'one coherent business capability',
+    ]);
+    expectAllNormalized(codeWriting, [
+      'Always read `references/repo-code-style.md`',
+      'For authoritative database or realtime service mutations, also read',
+      '`references/convergent-service-writing.md` completely',
+    ]);
+    for (const skillName of codeProducingSpecialistSkills) {
+      expect(
+        readRepo(`.agents/skills/${skillName}/SKILL.md`),
+        `${skillName} must route TypeScript work through rallar-code-writing`,
+      ).toContain('**REQUIRED SUB-SKILL:** Use `rallar-code-writing`');
+    }
+    for (const source of [testing, testCommands]) {
+      expect(source).toContain(canonicalServiceWritingPath);
+    }
+    expectAllNormalized(serviceWriting, [
+      'functional core',
+      'explicitly owned stateful shell',
+      'one coherent business capability, one ownership boundary, and one reason to change',
+      'records timing around each direct phase and around the AppInbox transaction separately',
+      'apply',
+      'no-op',
+      'reject',
+      'written',
+      'conflict',
+      'newer revisions',
+      'stale revisions',
+      'equal revisions with different canonical content',
+    ]);
+  });
+
+  it('requires code-derived authoritative mutation traces and canonical realtime owners', () => {
+    const codeWriting = readRepo('.agents/skills/rallar-code-writing/SKILL.md');
+    const codeStyle = readRepo('.agents/skills/rallar-code-writing/references/repo-code-style.md');
+    const serviceWriting = readRepo(canonicalServiceWritingPath);
+    const realtime = readRepo('.agents/skills/rallar-realtime/SKILL.md');
+
+    for (const source of [codeWriting, codeStyle, serviceWriting]) {
+      expectAllNormalized(source, [
+        'family-level code-derived trace',
+        'registration from invocation',
+        'external or protocol entry',
+        'callback registration owner and registration time',
+        'runtime invoker and callback invocation count or retry rule',
+        'representation translation and read, compute, validate, and write owners',
+        'transaction and retry owner and the first conditional guard',
+        'receipt, event, exact durable result, and final outbox writes',
+        'commit-return point and private after-commit data',
+        'after-commit effects, early exits, failures, and cleanup',
+        'final caller-visible result and canonical versus compatibility paths',
+        'mutable values do not escape a transaction callback',
+      ]);
+    }
+
+    expectAllNormalized(realtime, [
+      'group-state/**',
+      'topology/inbox/**',
+      'rtc-topology/inbox/**',
+      'services/**',
+      'compatibility-only',
+    ]);
+  });
+
+  it('requires review-pressure evidence before publishing a large implementation plan', () => {
+    const progress = readRepo('.agents/skills/publishing-plan-progress/SKILL.md');
+
+    expectAllNormalized(progress, [
+      'more than 100 changed files',
+      'more than 10,000 changed lines',
+      'more than 20 changed production modules',
+      'more than three materially different control-flow families',
+      'stacked-versus-single decision',
+      'one-screen read-first map',
+      'current head/tree/workflow evidence',
+      'stale evidence blocks completion',
+    ]);
+  });
+
+  it('keeps detailed ingress retry policy in the canonical service reference', () => {
+    const serviceWriting = readRepo(canonicalServiceWritingPath);
+    const routedSkillPaths = [
+      'AGENTS.md',
+      '.agents/skills/rallar-code-writing/SKILL.md',
+      '.agents/skills/rallar-code-writing/references/repo-code-style.md',
+      '.agents/skills/rallar-platform/SKILL.md',
+      '.agents/skills/rallar-realtime/SKILL.md',
+      '.agents/skills/rallar-testing/SKILL.md',
+      '.agents/skills/rallar-testing/references/test-commands.md',
+    ] as const;
+
+    expect(serviceWriting).toContain('1, 2, 4, 8, and 16 ms');
+    for (const filePath of routedSkillPaths) {
+      expect(readRepo(filePath), `${filePath} duplicates the retry schedule`).not.toContain(
+        '1, 2, 4, 8, and 16 ms',
+      );
+      expect(readRepo(filePath), `${filePath} must route to the canonical doctrine`).toContain(
+        canonicalServiceWritingPath,
+      );
+    }
+  });
+
+  it.each(currentAppInboxGuidancePaths)(
+    '%s states the current AppInbox transaction doctrine without stale precedent',
+    (filePath) => {
+      const guidance = normalizeWhitespace(readRepo(filePath));
+      expectAllNormalized(guidance, appInboxGuidanceVocabulary);
+      for (const rejected of [
+        'write opens the transaction',
+        '[0, 2, 8]',
+        'StateMutationOutboxWork',
+        'StateMutationOutboxRepository',
+        'commitTopologyWithRetry',
+      ]) {
+        expect(guidance, `${filePath}: ${rejected}`).not.toContain(rejected);
+      }
+      expect(guidance, filePath).not.toMatch(
+        /(?:the )?(?:`read`|read),\s*(?:`compute`|compute),\s*and\s*(?:`validate`|validate)\s+phases are pure|pure\s+(?:`read`|read)[,/]\s*(?:`compute`|compute)[,/]\s*(?:and\s*)?(?:`validate`|validate)\s+(?:phases|stages)/i,
+      );
+    },
+  );
+
+  it.each(postCommitAudienceGuidancePaths)(
+    '%s keeps logical WebSocket audience resolution after commit',
+    (filePath) => {
+      expect(normalizeWhitespace(readRepo(filePath))).toMatch(
+        /logical WebSocket audience resolution .{0,30}(?:only )?after commit/i,
+      );
+    },
+  );
+
+  it.each(repositoryReadGuidancePaths)(
+    '%s keeps repository reads outside the write transaction',
+    (filePath) => {
+      expectAllNormalized(readRepo(filePath), [
+        'read',
+        'loads the repository decision surface outside the write transaction',
+        'Only `compute` and `validate` are pure',
+      ]);
+    },
+  );
+
+  it.each(downstreamQueueGuidancePaths)(
+    '%s distinguishes downstream QueueBox retries from AppInbox ingress retries',
+    (filePath) => {
+      expectAllNormalized(readRepo(filePath), [
+        'RtcTopologyOutboxWork',
+        'ResourceInbox/QueueBox attempt boundary',
+        'neither service owns the transaction or retry boundary',
+      ]);
+    },
+  );
+
+  it.each(coreConvergentWriteGuidancePaths)(
+    '%s independently states the convergent-write doctrine',
+    (filePath) => {
+      const guidance = normalizeWhitespace(readRepo(filePath));
+      expect(guidance).toMatch(
+        /`compute` and `validate` .{0,30}(?:phases|functions) are .{0,10}pure/i,
+      );
+      expect(guidance).toMatch(/AppInbox .{0,50}owns .{0,40}transaction .{0,40}retry boundary/i);
+      expect(guidance).toMatch(/service write receives the transaction .{0,40}never opens/i);
+      expect(guidance).toContain('write(transaction, computed)');
+      expect(guidance).toMatch(
+        /(?:ResourceInboxRepository.{0,160}same transaction|same transaction.{0,160}ResourceInboxRepository)/i,
+      );
+      expect(guidance).toMatch(/queue locks.{0,80}coordination-only/i);
+      expect(guidance).toMatch(/computed persistence data.{0,40}not.{0,20}(?:called )?a plan/i);
+    },
+  );
+
+  it.each(canonicalSnapshotOrderingGuidancePaths)(
+    '%s requires canonical ordering for authoritative snapshot collections',
+    (filePath) => {
+      expectAllNormalized(readRepo(filePath), [
+        'Authoritative snapshot collections that represent unordered sets',
+        'canonical storage-key order',
+        'computed mutation result',
+        'durable repository assembly',
+        'arrival, insertion, or database/provider iteration order',
+        'equal-revision content checks',
+      ]);
+    },
+  );
+});
+
+function readRepo(filePath: string): string {
+  return readAbsolute(path.join(repoRoot, filePath));
+}
+
+function readAbsolute(filePath: string): string {
+  return readFileSync(filePath, 'utf8');
+}
+
+function readJson(filePath: string): unknown {
+  return JSON.parse(readRepo(filePath));
+}
+
+function expectAll(haystack: string, needles: readonly string[]): void {
+  for (const needle of needles) {
+    expect(haystack, needle).toContain(needle);
+  }
+}
+
+function expectAllNormalized(haystack: string, needles: readonly string[]): void {
+  const normalized = normalizeWhitespace(haystack);
+  for (const needle of needles) {
+    expect(normalized, needle).toContain(normalizeWhitespace(needle));
+  }
+}
+
+function normalizeWhitespace(value: string): string {
+  return value.replace(/\s+/g, ' ').trim();
+}
