+++ packages/tests/repo/rallar-skill-plugin-publication-integrity.test.ts
@@ -1,0 +1,224 @@
+import { existsSync, readFileSync, readdirSync } from 'node:fs';
+import path from 'node:path';
+import { describe, expect, it } from 'vitest';
+
+const repoRoot = process.cwd();
+const skillsRoot = path.join(repoRoot, '.agents/skills');
+const expectedSkills = [
+  'building-rallar-apps',
+  'performance-analysis',
+  'rallar-ai',
+  'rallar-code-writing',
+  'rallar-games',
+  'rallar-hetzner-ops',
+  'rallar-platform',
+  'publishing-plan-progress',
+  'rallar-realtime',
+  'rallar-testing',
+] as const;
+
+describe('Rallar skill plugin and publication integrity', () => {
+  it('uses one directly discoverable skill tree for the plugin', () => {
+    const plugin = readJson('.codex-plugin/plugin.json') as {
+      skills?: string;
+      interface?: { defaultPrompt?: readonly string[] };
+    };
+    const skillDirectories = readdirSync(skillsRoot, { withFileTypes: true })
+      .filter((entry) => entry.isDirectory())
+      .map((entry) => entry.name)
+      .sort();
+
+    expect(plugin.skills).toBe('./.agents/skills/');
+    expect(plugin.interface?.defaultPrompt?.length).toBeLessThanOrEqual(3);
+    expect(skillDirectories).toEqual([...expectedSkills].sort());
+    expect(existsSync(path.join(repoRoot, 'skills'))).toBe(false);
+  });
+
+  it('keeps skill frontmatter and local references valid', () => {
+    for (const skillName of expectedSkills) {
+      const skillPath = path.join(skillsRoot, skillName, 'SKILL.md');
+      const source = readAbsolute(skillPath);
+      const frontmatter = readFrontmatter(source, skillPath);
+
+      expect(frontmatter.name, skillPath).toBe(skillName);
+      expect(frontmatter.description.length, skillPath).toBeGreaterThan(20);
+      expect(frontmatter.description, skillPath).toMatch(/^Use when\b/);
+
+      for (const reference of source.matchAll(/`(references\/[a-z0-9./-]+\.md)`/g)) {
+        expect(
+          existsSync(path.join(skillsRoot, skillName, reference[1])),
+          `${skillPath} -> ${reference[1]}`,
+        ).toBe(true);
+      }
+    }
+  });
+
+  it('routes long-running written-plan execution to observable published progress', () => {
+    const agents = readRepo('AGENTS.md');
+    const progressSkill = readRepo('.agents/skills/publishing-plan-progress/SKILL.md');
+    const normalizedAgents = normalizeWhitespace(agents);
+    const normalizedProgressSkill = normalizeWhitespace(progressSkill);
+    const plugin = readJson('.codex-plugin/plugin.json') as {
+      keywords?: readonly string[];
+      interface?: { defaultPrompt?: readonly string[]; longDescription?: string };
+    };
+
+    expectAll(agents, ['publishing-plan-progress', 'long-running']);
+    expect(normalizedAgents).toContain(
+      'For written implementation plans and clearly long-running repository implementation, ' +
+        'including docs, scripts, and operations, use `publishing-plan-progress`.',
+    );
+    expectAll(normalizedAgents, [
+      'No AI or agent may create or place a commit on `main`, `master`, or the local default branch',
+      'permission immediately before the commit',
+      'Each default-branch commit requires a new permission request and approval',
+      'No AI or agent may push `main`, `master`, or the remote default branch',
+      'the push is forced',
+      'permission immediately before the push',
+      'Each default-branch push requires a new permission request and approval',
+      'Commit and push permissions are independent',
+    ]);
+    expectAll(normalizedProgressSkill, [
+      '`codex/<topic>`',
+      'draft pull request',
+      'On a non-default branch',
+      'completed plan milestone',
+      'without waiting for human review',
+      'Explicit user instructions',
+      'installed GitHub publication workflow',
+      'upstream tracking',
+      'detached worktree',
+      'Create branch here',
+      'Explicit user instructions may narrow or disable publication',
+      'Default Branch Commit and Push Permission',
+      'Before every default-branch commit',
+      'staged diff summary and staged Git tree ID',
+      '`git write-tree`',
+      'full commit IDs',
+      'proposed commit message, and operation type',
+      'commit, amend, merge, revert, cherry-pick, rebase, or squash',
+      'Do not commit while the user is unavailable or treat silence as approval',
+      'Any later default-branch commit requires a new request and approval',
+      'content, message, input commit, conflict resolution, or target changes',
+      'Working directly on the default branch is not permission to commit',
+      'Approval to commit is not approval to push',
+      'approval to push is not approval to commit',
+      'standing instruction to publish progress',
+      'exact remote, destination ref and refspec',
+      'resolved full commit IDs, not moving symbolic ranges',
+      'commit range changes, the approval is invalid',
+      'Wait for explicit approval',
+      'treat silence as approval',
+      'Any later default-branch push requires a new request and approval.',
+      'force push was separately disclosed',
+      'any source branch is pushed to the remote default ref',
+      'pauses only that Git operation, not safe uncommitted local plan work',
+    ]);
+    expect(plugin.keywords).toEqual(
+      expect.arrayContaining([
+        'implementation-plans',
+        'observable-plan-progress',
+        'draft-pull-requests',
+      ]),
+    );
+    expect(plugin.interface?.longDescription).toContain('observable plan progress');
+    expect(plugin.interface?.defaultPrompt).toHaveLength(3);
+    expect(plugin.interface?.defaultPrompt).toContain(
+      'Execute a long-running Rallar implementation plan with observable GitHub checkpoints.',
+    );
+  });
+
+  it('keeps plan completion behind local and published release gates', () => {
+    const agents = readRepo('AGENTS.md');
+    const progressSkill = readRepo('.agents/skills/publishing-plan-progress/SKILL.md');
+    const testingSkill = readRepo('.agents/skills/rallar-testing/SKILL.md');
+    const testCommands = readRepo('.agents/skills/rallar-testing/references/test-commands.md');
+    const codeWritingSkill = readRepo('.agents/skills/rallar-code-writing/SKILL.md');
+
+    for (const source of [agents, progressSkill, testingSkill, testCommands]) {
+      expectAllNormalized(source, [
+        'npm run test:unit',
+        'npm run test:ci',
+        'npm run build',
+        'draft pull request',
+        'Branch Release Gate',
+        'Run Hetzner Supported Distributed Manifests',
+        'final feature-branch commit',
+        'resulting default-branch commit',
+        'pending, skipped, failed, or attached to an older commit',
+        'the plan is not complete',
+      ]);
+    }
+
+    expectAllNormalized(progressSkill, [
+      'An instruction not to commit or push postpones publication',
+      'does not waive any completion gate',
+      'Any change after a successful gate invalidates that gate',
+    ]);
+    expectAllNormalized(testingSkill, [
+      'Focused tests are feedback, not a substitute for completion gates',
+      'Run the commands from the final uncommitted working tree before publication',
+    ]);
+    expectAllNormalized(testCommands, ['Plan Completion Gate', 'Release Gate', 'exact commit SHA']);
+    expectAllNormalized(codeWritingSkill, [
+      'rallar-testing',
+      'plan-completion gate',
+      'Focused checks never substitute for that final gate',
+    ]);
+  });
+
+  it('requires a collapsed learning-oriented command summary in every final handoff', () => {
+    const agents = readRepo('AGENTS.md');
+
+    expectAllNormalized(agents, [
+      'Every final handoff ends with a collapsed `<details>` block',
+      '<summary>Commands executed and what they taught us</summary>',
+      'If no commands or tool actions ran, say so inside the collapsed block',
+      'Group repeated or equivalent commands',
+      'why the command or action was chosen',
+      'what its result means',
+      'useful lesson or reusable troubleshooting insight',
+      'Never expose secrets, tokens, credentials, authorization headers',
+    ]);
+  });
+});
+
+function readRepo(filePath: string): string {
+  return readAbsolute(path.join(repoRoot, filePath));
+}
+
+function readAbsolute(filePath: string): string {
+  return readFileSync(filePath, 'utf8');
+}
+
+function readJson(filePath: string): unknown {
+  return JSON.parse(readRepo(filePath));
+}
+
+function expectAll(haystack: string, needles: readonly string[]): void {
+  for (const needle of needles) {
+    expect(haystack, needle).toContain(needle);
+  }
+}
+
+function expectAllNormalized(haystack: string, needles: readonly string[]): void {
+  const normalized = normalizeWhitespace(haystack);
+  for (const needle of needles) {
+    expect(normalized, needle).toContain(normalizeWhitespace(needle));
+  }
+}
+
+function normalizeWhitespace(value: string): string {
+  return value.replace(/\s+/g, ' ').trim();
+}
+
+function readFrontmatter(
+  source: string,
+  filePath: string,
+): Readonly<{ name: string; description: string }> {
+  const block = source.match(/^---\n([\s\S]*?)\n---(?:\n|$)/)?.[1];
+  expect(block, filePath).toBeDefined();
+  const name = block?.match(/^name:\s*(.+)$/m)?.[1]?.trim() ?? '';
+  const description = block?.match(/^description:\s*(.+)$/m)?.[1]?.trim() ?? '';
+  return { name, description };
+}
